SSD MobileNet v1 small fixes
* evaluation now run with single eval.py script
* add pretrained weights to the index
