--- ssd_mobilenetv1/trained_models.json
@@ -1,28 +1,0 @@
-{
-    "id": "ssdmobilenetv1",
-    "full_name": "SSD MobileNetV1",
-    "variants": [
-        {
-            "id": "baseline-keras",
-            "compressed": false,
-            "weights_url": "tba",
-            "training_dataset_id": null,
-            "training_dataset_variant_id": null,
-            "repo_url": "https://github.com/latentai/model-zoo-models/",
-            "model_schema": {
-                "output_names": "predictions/concat",
-                "input_names": "input_1",
-                "input_shapes": "1, 300, 300, 3"
-            },
-            "compress": {
-                "type": null,
-                "bit_depth": null,
-                "quantizer": null,
-                "skip_layers": null
-                },
-            "compile": {
-                "type": null
-            }
-        }
-    ]
-}
